requirements.txt
streamlit
openai-whisper
pandas
pytube
openai
openpyxl
ffmpeg-python


